updating stock images and positioned them
