--- src/components/future/carousel/carousel.css
@@ -1,19 +1,0 @@
-.carousel-container {
-    width: 100%;
-    max-width: 900px; /* Adjust the max-width as needed */
-    margin: 0 auto;
-}
-.carousel-item img {
-    width: 100%;
-    height: auto;
-    max-height: 500px; /* Adjust the max-height as needed */
-    object-fit: cover;
-}
-
-/* Media query for phones */
-@media (max-width: 600px) {
-    .carousel-container img {
-        width: 100%; /* Full width for smaller screens */
-        height: auto; /* Maintain the aspect ratio */
-    }
-}
